% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emmaneurosci-package.R
\docType{package}
\name{emmaneurosci-package}
\alias{emmaneurosci}
\alias{emmaneurosci-package}
\title{emmaneurosci: Emma Neuroscience Group functions (Emma Children's Hospital, Amsterdam UMC)}
\description{
ools for cognitive neuroscience, including cluster analysis,
neurocognitive network reconstruction, network analysis. Developed for
the Emma Neuroscience Group, Emma Children's Hospital, Amsterdam UMC.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mkonigs/emmaneurosci}
  \item Report bugs at \url{https://github.com/mkonigs/emmaneurosci}
}

}
\author{
\strong{Maintainer}: 

Authors:
\itemize{
  \item Marsh Königs \email{m.konigs@amsterdamumc.nl}
}

}
\keyword{internal}
